Added missing cast operators
git-svn-id: http://svn.php.net/repository/php/php-src/php/php-src/trunk@111297 c90b9560-bf6c-de11-be94-00142212c4b1

+++ ext/iconv/iconv.c
@@ -641,7 +641,7 @@ static php_iconv_err_t _php_iconv_substr(smart_str *pretval,
 			}
 		}
 
-		if (cnt >= offset) {
+		if (cnt >= (unsigned int)offset) {
 			if (cd2 == NULL) {
 				cd2 = icv_open(enc, GENERIC_SUPERSET_NAME);
 
@@ -790,7 +790,7 @@ static php_iconv_err_t _php_iconv_strpos(unsigned int *pretval,
 			}
 		}
 		if (offset >= 0) {
-			if (cnt >= offset) {
+			if (cnt >= (unsigned int)offset) {
 				if (_php_iconv_memequal(buf, ndl_buf_p, sizeof(buf))) {
 					if (match_ofs == (unsigned int)-1) {
 						match_ofs = cnt;
